fix: tab in tablet
